Make benchmarks more uniform
